front draggable と draggable しないモードを整備
