+++ memo/to-do.txt
@@ -16,13 +16,15 @@
 [済][4/29] id が1でない場合の単独コンテンツID特定バグ
 [済][4/30] 顧客データにテンプレート名追加
 [済][5/12] 画像の拡大移動をサポートする、データ構造も拡張する
+[済][5/12] front draggable と draggable しないモードを整備
 
-** front draggable と draggable しないモードを整備
+** contents-grid flex ではなく grid に変更
 ** nuxt auth expire 時間が短い問題
 
 ** メールとアイキャッチと別のコンテンツとの切り離し
 ** テーマ対応 まずは default (longlivenet)と dark (states向け) を分離
-** TAG と Reccomend 機能追加
+** TAG 機能追加
+** Reccomend 機能追加
 ** SNS 部品埋め込み
 ** HOME画面スクロール時に UP ボタン表示
 * コンテンツタイトル文字色の変更 カラーピッカー https://github.com/saintplay/vue-swatches
@@ -39,7 +41,8 @@
 
 ** 問い合わせメール送信部分の実装
 ** メールとアイキャッチと別のコンテンツとの切り離し
-** TAG と Reccomend 機能追加
+** TAG 機能追加
+** Reccomend 機能追加
 * 管理画面認証機能
 * 顧客ユーザ認証処理のpassportを使用した見直し(可能なら)
